requirements.txt
argparse==1.4.0
Cython==0.29.30
git+https://github.com/YyuK-Liao/cython_bbox.git#egg=cython-bbox
lap==0.4.0
loguru==0.6.0
numpy==1.23.1
opencv-python==4.5.5.64
pandas==1.4.3
pycocotools==2.0.4
PyYAML==6.0
scikit-image==0.19.3
scipy==1.8.1
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
thop==0.1.0.post2207010342
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.64.0
ujson==5.4.0
wheel==0.37.1
yacs==0.1.8
rich==12.5.1